REF: collect Timestamp tests by method
