+++ pandas/tests/scalar/timestamp/methods/test_round.py
@@ -1,6 +1,3 @@
-from datetime import datetime
-
-from dateutil.tz import gettz
 from hypothesis import (
     given,
     strategies as st,
@@ -8,7 +5,6 @@ from hypothesis import (
 import numpy as np
 import pytest
 import pytz
-from pytz import utc
 
 from pandas._libs import lib
 from pandas._libs.tslibs import (
@@ -16,19 +12,16 @@ from pandas._libs.tslibs import (
     OutOfBoundsDatetime,
     Timedelta,
     Timestamp,
-    conversion,
     iNaT,
     to_offset,
 )
 from pandas._libs.tslibs.dtypes import NpyDatetimeUnit
 from pandas._libs.tslibs.period import INVALID_FREQ_ERR_MSG
-import pandas.util._test_decorators as td
 
 import pandas._testing as tm
 
 
-class TestTimestampUnaryOps:
-    # --------------------------------------------------------------
+class TestTimestampRound:
     def test_round_division_by_zero_raises(self):
         ts = Timestamp("2016-01-01")
 
@@ -36,7 +29,6 @@ class TestTimestampUnaryOps:
         with pytest.raises(ValueError, match=msg):
             ts.round("0ns")
 
-    # Timestamp.round
     @pytest.mark.parametrize(
         "timestamp, freq, expected",
         [
@@ -389,221 +381,3 @@ class TestTimestampUnaryOps:
 
         nanos = 24 * 60 * 60 * 1_000_000_000
         checker(ts, nanos, "D")
-
-    # --------------------------------------------------------------
-    # Timestamp.replace
-
-    def test_replace_out_of_pydatetime_bounds(self):
-        # GH#50348
-        ts = Timestamp("2016-01-01").as_unit("ns")
-
-        msg = "Out of bounds nanosecond timestamp: 99999-01-01 00:00:00"
-        with pytest.raises(OutOfBoundsDatetime, match=msg):
-            ts.replace(year=99_999)
-
-        ts = ts.as_unit("ms")
-        result = ts.replace(year=99_999)
-        assert result.year == 99_999
-        assert result._value == Timestamp(np.datetime64("99999-01-01", "ms"))._value
-
-    def test_replace_non_nano(self):
-        ts = Timestamp._from_value_and_reso(
-            91514880000000000, NpyDatetimeUnit.NPY_FR_us.value, None
-        )
-        assert ts.to_pydatetime() == datetime(4869, 12, 28)
-
-        result = ts.replace(year=4900)
-        assert result._creso == ts._creso
-        assert result.to_pydatetime() == datetime(4900, 12, 28)
-
-    def test_replace_naive(self):
-        # GH#14621, GH#7825
-        ts = Timestamp("2016-01-01 09:00:00")
-        result = ts.replace(hour=0)
-        expected = Timestamp("2016-01-01 00:00:00")
-        assert result == expected
-
-    def test_replace_aware(self, tz_aware_fixture):
-        tz = tz_aware_fixture
-        # GH#14621, GH#7825
-        # replacing datetime components with and w/o presence of a timezone
-        ts = Timestamp("2016-01-01 09:00:00", tz=tz)
-        result = ts.replace(hour=0)
-        expected = Timestamp("2016-01-01 00:00:00", tz=tz)
-        assert result == expected
-
-    def test_replace_preserves_nanos(self, tz_aware_fixture):
-        tz = tz_aware_fixture
-        # GH#14621, GH#7825
-        ts = Timestamp("2016-01-01 09:00:00.000000123", tz=tz)
-        result = ts.replace(hour=0)
-        expected = Timestamp("2016-01-01 00:00:00.000000123", tz=tz)
-        assert result == expected
-
-    def test_replace_multiple(self, tz_aware_fixture):
-        tz = tz_aware_fixture
-        # GH#14621, GH#7825
-        # replacing datetime components with and w/o presence of a timezone
-        # test all
-        ts = Timestamp("2016-01-01 09:00:00.000000123", tz=tz)
-        result = ts.replace(
-            year=2015,
-            month=2,
-            day=2,
-            hour=0,
-            minute=5,
-            second=5,
-            microsecond=5,
-            nanosecond=5,
-        )
-        expected = Timestamp("2015-02-02 00:05:05.000005005", tz=tz)
-        assert result == expected
-
-    def test_replace_invalid_kwarg(self, tz_aware_fixture):
-        tz = tz_aware_fixture
-        # GH#14621, GH#7825
-        ts = Timestamp("2016-01-01 09:00:00.000000123", tz=tz)
-        msg = r"replace\(\) got an unexpected keyword argument"
-        with pytest.raises(TypeError, match=msg):
-            ts.replace(foo=5)
-
-    def test_replace_integer_args(self, tz_aware_fixture):
-        tz = tz_aware_fixture
-        # GH#14621, GH#7825
-        ts = Timestamp("2016-01-01 09:00:00.000000123", tz=tz)
-        msg = "value must be an integer, received <class 'float'> for hour"
-        with pytest.raises(ValueError, match=msg):
-            ts.replace(hour=0.1)
-
-    def test_replace_tzinfo_equiv_tz_localize_none(self):
-        # GH#14621, GH#7825
-        # assert conversion to naive is the same as replacing tzinfo with None
-        ts = Timestamp("2013-11-03 01:59:59.999999-0400", tz="US/Eastern")
-        assert ts.tz_localize(None) == ts.replace(tzinfo=None)
-
-    @td.skip_if_windows
-    def test_replace_tzinfo(self):
-        # GH#15683
-        dt = datetime(2016, 3, 27, 1)
-        tzinfo = pytz.timezone("CET").localize(dt, is_dst=False).tzinfo
-
-        result_dt = dt.replace(tzinfo=tzinfo)
-        result_pd = Timestamp(dt).replace(tzinfo=tzinfo)
-
-        # datetime.timestamp() converts in the local timezone
-        with tm.set_timezone("UTC"):
-            assert result_dt.timestamp() == result_pd.timestamp()
-
-        assert result_dt == result_pd
-        assert result_dt == result_pd.to_pydatetime()
-
-        result_dt = dt.replace(tzinfo=tzinfo).replace(tzinfo=None)
-        result_pd = Timestamp(dt).replace(tzinfo=tzinfo).replace(tzinfo=None)
-
-        # datetime.timestamp() converts in the local timezone
-        with tm.set_timezone("UTC"):
-            assert result_dt.timestamp() == result_pd.timestamp()
-
-        assert result_dt == result_pd
-        assert result_dt == result_pd.to_pydatetime()
-
-    @pytest.mark.parametrize(
-        "tz, normalize",
-        [
-            (pytz.timezone("US/Eastern"), lambda x: x.tzinfo.normalize(x)),
-            (gettz("US/Eastern"), lambda x: x),
-        ],
-    )
-    def test_replace_across_dst(self, tz, normalize):
-        # GH#18319 check that 1) timezone is correctly normalized and
-        # 2) that hour is not incorrectly changed by this normalization
-        ts_naive = Timestamp("2017-12-03 16:03:30")
-        ts_aware = conversion.localize_pydatetime(ts_naive, tz)
-
-        # Preliminary sanity-check
-        assert ts_aware == normalize(ts_aware)
-
-        # Replace across DST boundary
-        ts2 = ts_aware.replace(month=6)
-
-        # Check that `replace` preserves hour literal
-        assert (ts2.hour, ts2.minute) == (ts_aware.hour, ts_aware.minute)
-
-        # Check that post-replace object is appropriately normalized
-        ts2b = normalize(ts2)
-        assert ts2 == ts2b
-
-    @pytest.mark.parametrize("unit", ["ns", "us", "ms", "s"])
-    def test_replace_dst_border(self, unit):
-        # Gh 7825
-        t = Timestamp("2013-11-3", tz="America/Chicago").as_unit(unit)
-        result = t.replace(hour=3)
-        expected = Timestamp("2013-11-3 03:00:00", tz="America/Chicago")
-        assert result == expected
-        assert result._creso == getattr(NpyDatetimeUnit, f"NPY_FR_{unit}").value
-
-    @pytest.mark.parametrize("fold", [0, 1])
-    @pytest.mark.parametrize("tz", ["dateutil/Europe/London", "Europe/London"])
-    @pytest.mark.parametrize("unit", ["ns", "us", "ms", "s"])
-    def test_replace_dst_fold(self, fold, tz, unit):
-        # GH 25017
-        d = datetime(2019, 10, 27, 2, 30)
-        ts = Timestamp(d, tz=tz).as_unit(unit)
-        result = ts.replace(hour=1, fold=fold)
-        expected = Timestamp(datetime(2019, 10, 27, 1, 30)).tz_localize(
-            tz, ambiguous=not fold
-        )
-        assert result == expected
-        assert result._creso == getattr(NpyDatetimeUnit, f"NPY_FR_{unit}").value
-
-    # --------------------------------------------------------------
-    # Timestamp.normalize
-
-    @pytest.mark.parametrize("arg", ["2013-11-30", "2013-11-30 12:00:00"])
-    @pytest.mark.parametrize("unit", ["ns", "us", "ms", "s"])
-    def test_normalize(self, tz_naive_fixture, arg, unit):
-        tz = tz_naive_fixture
-        ts = Timestamp(arg, tz=tz).as_unit(unit)
-        result = ts.normalize()
-        expected = Timestamp("2013-11-30", tz=tz)
-        assert result == expected
-        assert result._creso == getattr(NpyDatetimeUnit, f"NPY_FR_{unit}").value
-
-    def test_normalize_pre_epoch_dates(self):
-        # GH: 36294
-        result = Timestamp("1969-01-01 09:00:00").normalize()
-        expected = Timestamp("1969-01-01 00:00:00")
-        assert result == expected
-
-    # --------------------------------------------------------------
-
-    @td.skip_if_windows
-    def test_timestamp(self, fixed_now_ts):
-        # GH#17329
-        # tz-naive --> treat it as if it were UTC for purposes of timestamp()
-        ts = fixed_now_ts
-        uts = ts.replace(tzinfo=utc)
-        assert ts.timestamp() == uts.timestamp()
-
-        tsc = Timestamp("2014-10-11 11:00:01.12345678", tz="US/Central")
-        utsc = tsc.tz_convert("UTC")
-
-        # utsc is a different representation of the same time
-        assert tsc.timestamp() == utsc.timestamp()
-
-        # datetime.timestamp() converts in the local timezone
-        with tm.set_timezone("UTC"):
-            # should agree with datetime.timestamp method
-            dt = ts.to_pydatetime()
-            assert dt.timestamp() == ts.timestamp()
-
-
-@pytest.mark.parametrize("fold", [0, 1])
-def test_replace_preserves_fold(fold):
-    # GH 37610. Check that replace preserves Timestamp fold property
-    tz = gettz("Europe/Moscow")
-
-    ts = Timestamp(year=2009, month=10, day=25, hour=2, minute=30, fold=fold, tzinfo=tz)
-    ts_replaced = ts.replace(second=1)
-
-    assert ts_replaced.fold == fold
